requirements.txt
Flask
gunicorn
python-pptx
requests